direct: Support configuration for unordered slices; use in permissions to ignore order (#4014)
## Changes
- Extend structpath to support [key=‘value’] syntax.
- Extend structdiff with extra parameter that maps path to keyFunc which
is a function that returns key/value pair for a given slice element.
They pair is used to identify the slice element. 
- Use this functionality on permissions resource in direct to solve
drift due to backend reordering permissions.

## Why
- Prevent remote drift due to backend reordering slice elements.
- Make diffs more readable / precise.

This is an alternative to always sorting the input, which destroys the
references ($resources.jobs.foo.tasks[1] is not valid today because of
sorting). In a follow up PR I’ll revert job tasks sorting
https://github.com/databricks/cli/pull/3212 and implement KeyedSlices
instead.