Fix `databricks labs list` to include all projects (#3935)
## Changes

This PR fixes the `databricks labs list` command so that all projects
from Databricks Labs are listed: prior to this PR only the first 30
projects are listed.

The root cause of the problem is that the GitHub API being used is
paginated, and defaults to only 30 results per page, and pagination was
not implemented. This PR resolves the issue by:

 - updating the implementation to use the pagination mechanism; and
- increasing the number of results per page from 30 (default) to 100
(max allowed).

Incidental changes include:

- Some trace-level logging of the requests that are being made to the
GitHub API, consistent with elsewhere in the codebase.
 - Some spelling fixes.

Out of scope for now:

- Filtering the list of projects to only cover those that can be
installed.

## Why

The `databricks labs list` command was returning an incomplete list;
projects that could be installed (eg. `ucx`) were not included in the
list.

## Tests

 - Additional test case to cover pagination of the GitHub API.
 - Manual testing.