DAB unable to validate container local file:///path-to-jar (#3892)
_##_ Changes
In a runtime container, I need to be able to reference files. 
This PR allows the `file:///` path (which is valid in databricks
templates) to do that.
Currently the CLI attempts to resolve this path when it should now. 

Open to feedback here.

## Why
When running a containerized job, I have pre-bundled assets that need to
be referenced.
If I set the path to `file:///` manually via the databricks UX, my job
works. If I attempt to do it via the cli, I have this rendering issue.

```
databricks bundle deploy --target dev
Error: file doesn't exist file:///opt/spark/jars/app.jar
```
With this change, the asset bundle is deployable. 

## Tests
Unit testing and manual deployment verification.

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->

Signed-off-by: sshpuntoff <samshpuntoff@gmail.com>