Add testmask tool for conditional CI test execution (#4017)
## Changes

A new tool `testmask` under `tools/` that determines the set of Makefile
targets to run for testing.

The test jobs all depend on a job that runs this tool and use its output
to determine if they should run or skip. If a change only touches files
under `experimental/ssh`, for example, it won't trigger the main test
target. We can extend this to add more selective targets, only if we are
sure that there is no dependency between the files matched for the
target and other files in the repository. This, for now, relies on
convention rather than strict enforcement.

### How

Every file in the diff between the PR and the PR base is attributed to
one of the targets. If all files map to the same target, only that
target is run. If files map to multiple targets, multiple targets are
run. If there are files that cannot be mapped to a specific target, then
the main `test` target is run.

### Mapping to GitHub Actions

This approach uses a GitHub Actions test job for every target. This
results in a status check mark for each of them, making it easy to
eyeball which tests ran on a PR, which tests can on main, and when they
broke. Executing the same locally is straightforward because they are
all named after Makefile targets.

It applies to pull requests and merge queue checks. Scheduled runs and
main branch pushes always run the full test suite.

## Why

Not all tests need to run for every change. Being more selective reduces
build times and removes noise in PRs.

## Tests

The tests for this PR show it works.

A new test in `tools/testmask` confirms that the targets it maps to are
present in the Makefile.

PRs for each of the targets:
* #4023 
* #4024
* #4025
* #4026