Include old & new values for changes in the plan (#4072)
## Why
Useful for debugging. Could also be useful for resource implementation
given that we pass PlanEntry to DoUpdate() now.

Note, "new" value is redundant because it comes from remote_state or
new_state that is already included in the plan. But it could be still
difficult for 3rd party tools to parse this out as it requires to
support semantics of structpath (which includes things like
`tasks[task_key='mytask']`).

On the other hand, "old" value is not redundant as we don't include
prior state in the plan at the moment.