Run check workflow on push to main to populate cache (#4075)
## Why

Only if a cache is stored from a run from the main branch will it be
picked up in PRs and the merge queue. After splitting out the lint job
into a dedicated workflow (#4013), it no longer ran on the main branch,
and we no longer had cache hits for the lint jobs on PRs and the merge
queue.