Add support for alerts to DABs (#4004)
## Changes
Support for alerts was removed until API side issues were fixed. This PR
enables alerts in DABs again.

I had to regenerate openapi annotations and the pydabs code because the
linter was complaining. It seems that we do not automatically update
these.

## Tests
Now the acceptance test is also run on cloud.