Group test results in a single downstream job (#4078)
## Why

We need a single "required check" for tests. Now that we split tests
into separate jobs that may or may not be skipped, they can no longer be
required checks themselves. This new downstream job that groups the
result of all of them will always run and can become the new "required
check".