Include version in state and plan (#4126)
## Changes
- Include state_version=1 in the state.
- Include plan_version=1 in the plan.
- Include cli_version in both.

## Why
Good for debugging. In the future we might want to introduce version 2,
this would help.

We'll have to support reading state without the version info though. We
don't yet read plans, so once we start doing it
(https://github.com/databricks/cli/issues/4094), we can limit ourselves
to plan_version==1.