direct: parse state via json.RawMessage into appropriate type (#4138)
## Changes
Instead of parsing into any and then converting into the target type, we
now store json.RawMessage and then parse into target type when needed.

## Why
One issue it fixes is that JSON parser by default parses integers as
float64, losing precision. This was found by @varundeepsaini and a fix
is proposed in https://github.com/databricks/cli/pull/4136

This PR is more general though, fully relying on the type to parse
itself and it should more efficient as well as we're not walking the
type twice & we're not having 2 copies of it.

## Tests
New test where int64 field is initialized with large numbers. It fails
on main.