fix: dev-remote hmr port override (#4163)
## Changes
<!-- Brief summary of your changes that is easy to understand -->
This removes the ability to override the hmr port in the vite config to
avoid breaking the system.

## Why
<!-- Why are these changes needed? Provide the context that the reviewer
might be missing.
For example, were there any decisions behind the change that are not
reflected in the code itself? -->

vite uses the same port for the server and hmr, so by just providing the
port override via flag which we already have it works fine.
The problem comes when a user overrides the vite.config.js file with
this

```
  server: {
    hmr: {
      port: 1234,
    },
  },
```

This breaks the system, and there's no reason the user will need that
when running the dev-remote command as we control it

## Tests
<!-- How have you tested the changes? -->

<!-- If your PR needs to be included in the release notes for next
release,
add a separate entry in NEXT_CHANGELOG.md as part of your PR. -->

Co-authored-by: MarioCadenas <MarioCadenas@users.noreply.github.com>