Fix deployment failing when app is in DELETING state (#4102)
Fixes: #4006

## Changes
Added a wait for apps in `DELETING` state to be fully deleted before
deployment proceeds.

## Why
Deployment fails when an app is in `DELETING` state because the app
still exists but is being removed. This adds a pre-deployment check that
polls and waits (up to 5 minutes) for apps to finish deleting.