Fix incorrect test selection when -update if present
If -update flag is present, the test runner does not select the right
subtest but picks the first one always. This made sense when all subtests
were expected to generate the same output but that's not the case
anymore. There is attempt to fix it by checking EnvVaryOutput but that
just makes it more confusing.

Also, when debugging a test, it could be useful to record output of
specific subtest as a comparison base.